+++ docs/internals/architecture/harness/coding-shared-layer-owner-rebaseline.md
@@ -45,7 +45,7 @@ remain behind injected AI/Product ports.
 | Source region | LOC | Current shared owner or adopted mechanism | Classification | Next action |
 | --- | ---: | --- | --- | --- |
 | removed `coding.extensions` | 0 | `harness.extensions.agent` over the existing neutral extension loader/runtime | `shared adopted` | Standard Agent API, policy, loader, and runner profile moved without retaining a Product facade |
-| `coding.event.*` | 1,045 -> ~35 canonical facade | `harness.events.session_types`, `harness.events.session_projection`, `harness.events.runtime_views`, and `harness.events.recording_policy` own shared contracts, standard views, render enrichment, stream shaping, snake_case serialization, runtime-view selection, delivery hints, transcript-write decisions, and cancellation classification | `product adapter` | Keep only the established event facade, Product/Work mapping, and final presentation; the moved runtime-view and recording-policy modules must not grow Coding logic |
+| removed `coding.event` | 1,045 -> 0 | `harness.events.session_types`, `harness.events.session_projection`, `harness.events.runtime_views`, and `harness.events.recording_policy` own shared contracts, standard views, render enrichment, stream shaping, snake_case serialization, runtime-view selection, delivery hints, transcript-write decisions, and cancellation classification | `shared adopted` | Consumers import the canonical event owners directly; Product/Work mapping and final presentation remain in their existing owners without a Coding event facade |
 | removed `coding.control.settings_manager` | 0 | `harness.config.agent.SettingsManager` composed over `SettingsRuntime`, `ScopedConfigRuntime`, schema codec, and JSON store | `shared adopted` | Standard Agent field codecs, accessors, and mutations moved without adding a second engine |
 | removed `coding.control.types` | 0 | `harness.config.agent.types` | `shared adopted` | Standard Agent settings records are shared; Products retain only true domain additions and overlays |
 | `coding.policy.*` | 515 | Harness rule, approval, and resource-policy mechanisms | `product adapter` + `product kernel` | Extract profiles, not rules mechanically |
