--- src/loushang/coding/control/model_registry.py
@@ -1,5 +1,0 @@
-"""Public Coding import for the shared AI model catalog."""
-
-from loushang.harness.model_catalog import ModelCatalog as ModelRegistry
-
-__all__ = ["ModelRegistry"]
