--- src/loushang/coding/prompt/preflight.py
@@ -1,17 +1,0 @@
-"""Compatibility exports for Harness prompt preflight."""
-
-from loushang.harness.capabilities.prompt_preflight import (
-    PromptPreflightResult as PromptPreflightResult,
-)
-from loushang.harness.capabilities.prompt_preflight import (
-    preflight_user_input as preflight_user_input,
-)
-from loushang.harness.capabilities.prompt_preflight import (
-    preflight_user_input_async as preflight_user_input_async,
-)
-
-__all__ = [
-    "PromptPreflightResult",
-    "preflight_user_input",
-    "preflight_user_input_async",
-]
